+++ pybop/parameters/parameter.py
@@ -189,8 +189,15 @@ class Parameter:
         Return the initial value of each parameter.
         """
         if self.initial_value is None:
-            sample = self.rvs(1)
-            self.update(initial_value=sample[0])
+            if self.prior is not None:
+                sample = self.rvs(1)[0]
+                self.update(initial_value=sample)
+            else:
+                warnings.warn(
+                    "Initial value or Prior are None, proceeding without initial value.",
+                    UserWarning,
+                    stacklevel=2,
+                )
 
         return self.initial_value
 
@@ -409,17 +416,8 @@ class Parameters:
         initial_values = []
 
         for param in self.param.values():
-            if param.initial_value is None:
-                if param.prior is not None:
-                    initial_value = param.rvs(1)[0]
-                    param.update(initial_value=initial_value)
-                else:
-                    warnings.warn(
-                        "Initial value or Prior are None, proceeding without initial value.",
-                        UserWarning,
-                        stacklevel=2,
-                    )
-            initial_values.append(param.initial_value)
+            initial_value = param.get_initial_value()
+            initial_values.append(initial_value)
 
         return np.asarray(initial_values)